% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fig_label}
\alias{fig_label}
\title{Adding a text in a corner of a plot}
\usage{
fig_label(text, region = "figure", pos = "topleft", cex = NULL, ...)
}
\arguments{
\item{text}{string, the text to add to the plot}

\item{region}{string, the region where put the text on th eplot}

\item{pos}{interger, relative position of the letter}

\item{cex}{interger, size of the text}

\item{...}{Parameters passed to graphics::text function}
}
\description{
This function add a text in a corner of a plot
}
